job-tracker-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
gspread==5.12.3
oauth2client==4.1.3
groq==0.4.2
pydantic==2.5.3
httpx==0.26.0
beautifulsoup4==4.12.3
python-multipart==0.0.6
google-search-results==2.4.2
PyPDF2==3.0.1
